QPacketProtocol: Remove aboutToClose() method
It doesn't serve any useful purpose. The cleanup is done on destruction
anyway, and re-opening the underlying device after it was closed isn't
supported.

Furthermore, clearing the sendingPackets leads to a crash when the
device sends another bytesWritten() signal after aboutToClose(). That
can happen, for example with local sockets on windows, when flushing the
remaining data.

Fixes: QTBUG-70638
Change-Id: Icd636efc9794a127a3869a1c3e999d2286593067
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>