QQuickTableView: only preload items when rebuilding the table from scratch
We preload delegate items when the table is built from scratch to ensure
that we have enough items in circulation before the user starts to flick.
But when we only rebuild a section of the table after e.g adding or removing
rows, we already have enough items in circulation. So only preload when
building for RebuildOption::All.

Change-Id: I96aa9518e8032ecaba25f92fe6a0fbf17ea34c7d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>