Add QQuickItemPrivate::updateOrAddItemChangeListener()
This is identical to the existing updateOrAddGeometryChangeListener(),
except that it updates the regular "types" member, not "gTypes".

This function will allow preventing duplicate change listeners, at the
expensive of a call to QVector::indexOf(). It's useful when there is
no other way for calling code to check if a listener will be a
duplicate before adding it.

Task-number: QTBUG-69056
Task-number: QTBUG-70729
Change-Id: Idba039f355023e8d45a8b46e4af95aa81c13c3f4
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>