Software Adaptation: Improve bounding rect claculation for glyph nodes
It is not enough to call QGlyphRun::boundingRect() to determine the true
bounding rect of a series of glyphs. This issue would manifest itself
when rendering italic text.  Now we calculate the bounding rect for a
QGlyphRun when they are changed in the node so that that the clip rect
for glyphnodes reflect what is set on the node.

Task-number: QTBUG-68085
Change-Id: I834a7312052c8c5a4ad62f28333108c051d7f7b8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>