YarrJIT: disable MatchResult decoding constructor on 64bit
The YarrJIT generated code returns result differently on 32bit and 64bit
platforms: on 64bit the struct is returned (where Win64 is explicitly
handled separately), but on 32bit a uint64_t is returned. This
constructor takes care of decoding that uint64_t value, so it is only
needed on 32bit platforms. The code is actually invalid on 64bit.

Task-number: QTBUG-69996
Change-Id: Ie8cc737da55633aafa4bd6b3ebb8625b4ab354c1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>