Amend TouchMouse::buttonOnDelayedPressFlickable to test delayed tap
This test is for Flickable with pressDelay set to a testable timeout
value.  For some reason it only checked trying to flick before the
press delay is over, or holding and waiting for the press delay to expire.
But if the user taps a child MouseArea before the press delay expires
(neither moving nor waiting), at the moment of release it's supposed to
send the delayed press and then the release, so that the MouseArea
gets "clicked" almost as if there was no delay.  Lack of test coverage
allowed a regression in this functionality for 2 minor versions:
5.10 and 5.11.

Task-number: QTBUG-61144
Task-number: QTBUG-69059
Change-Id: I0d8867587e6877359c41ec5fc3a5cdd17447d0b8
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>