Examples: Use XmlHttpRequest to fetch flickr public photos
This change is to facilite moving the XmlListModel code to the
qtxmlpatterns repo, but still keep all the Particles examples
together.

Also fixed:
- the old TODO about darkening images according to depth,
and make it generally look a bit better.

- move the visualdatamodel examples to a directory called
delegatemodel, since that's how they work now.  And add them
to the main "views" example launcher so that they become more
visible to users.

Change-Id: I78f9f19e1e110608580adedcf4cd3be554222515
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>