Qml Tooling: Correctly sort debug messages into traces
Debug messages are sent from a different debug service and generally
arrive before they are "due". Therefore we buffer them and pass them on
when the trace has advanced to their position. However, we forgot some
places where we needed to check if there are pending debug messages to
be sent.

Change-Id: Ia6263086a9d06ae7c77c9e2afa85e6e1d8c8a19b
Reviewed-by: Michael Brasser <michael.brasser@live.com>