TapHandler: clean up when wantsEventPoint returns false
We already emitted grabCanceled() to inform QML callbacks, but we didn't
call reset().  It seems more proper to do everything that would normally
be done when grab is canceled.  TapHandler should not give up its passive
grab yet though, because that prevents delivery to any parent Flickable
that might be filtering events.  A parent Flickable should be able to
start flicking after the drag threshold is exceeded (it happens to be
exactly when TapHandler gives up).

Fixes: QTBUG-71466
Fixes: QTBUG-71970
Change-Id: Ibba1b0de92cfd88547eeb44edb095d019de76a94
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>