Doc: Rename landing page title to 'Qt Quick Controls'
Qt Quick Controls 2 module is referred to without the version number.
The landing page title still had the version number, which caused
the table of contents to fail to find the html filename.

Fix the landing page title and use the old title as a \keyword,
so any external links still continue to work.

Task-number: QTBUG-71694
Change-Id: I99e5eabf56028bd8a3180cb7161a0b0dcbdf9863
Reviewed-by: Martin Smith <martin.smith@qt.io>