QQuickScrollView: override getContentWidth/Height()
A ScrollView lets you add the Flickable that should be decorated as
a child. If that flickable has one (and only one) child, the
contentWidth/Height properties of the ScrollView will be calculated
by using the implicit size of that child (unless you assign something
else to those properties explicitly).

This logic goes wrong when the flickable is a TableView for several
reasons. The first is that TableView will populate the content item
dynamically, and for the first delegate item added, the content size
of the ScrollView will be set to be the size of this item (since the
TableView only got one child at that point). The second is that
TableView has its own set of contentWidth/Height properties. And those
properties are not respected by ScrollView. So even if TableView set
the contentWidth/Height to be the size of the table, this will not
be used by ScrollView. The result is that ScrollView concludes that
the content item is empty, which means that no scrollbars end up
visible or usable.

This patch will fix this by overriding getContentWidth()/Height()
from QQuickPane. The implementation will check if the flickable
has valid contentWidth/Height values set, and if so, use
them. Otherwise it will fall back to the old QQuickPane implementation
(which will inspect the children etc).

Fixes: QTBUG-71974
Change-Id: I027b9b939a10df2aeb816dea596adcb452f914b9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: J-P Nurmi <jpnurmi@gmail.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>