V4 tracing: Add a scheduler for the tracing JIT IR
This is a scheduler for the graph nodes, which first reconstructs the
control-flow graph, and then places all remaining nodes inside the basic
blocks.

The output of this pass is an MIFunction (MI = Machine Interface),
which uses a representation suitable for feeding to an assembler. Note
however that it still uses "virtual registers" at this point, so the
next pass will have to place those virtual registers in physical
registers or on a stack.

The code for the dominator tree calculation, block scheduling, loop info
and the blockset were lifted from the 5.10 JIT.

Change-Id: I11c4cc3f64fedba6dd4275b35bbea85d30d76f7d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>