Fix stale QML cache files when using assets:// on Android
When the "source file system" cannot provide us with a reliable time
stamp for a .qml file, do not try to save a cache file to disk.
Otherwise we may end up with stale cache files that are loaded despite
the source having changed.

We used to fall back to the app binary time stamp, but that was needed
before we correctly stored the time stamp of resources in the qrc data.

Fixes: QTBUG-68736
Change-Id: I8c2f8939d253472ce8bc1e3a59e8142d21f07539
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Richard Weickelt <richard@weickelt.de>