QML: Also clear outerContext's contextObject on destruction
A QObject can not only be set as contextObject of the own context, but
also as contextObject of the outerContext of a respective QQmlData. This
can be seen in QQmlObjectCreator::createInstance(...) if isContextObject
is true.

Therefore, when catching a QObject deletion we need to clear the pointer
in the outerContext if that refers to the object being deleted.

Fixes: QTBUG-71037
Change-Id: Ib6ba99bd5336f7582486b2128515021245370c60
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>