SplitView: schedule another layout if requested sizes changed during layout
As mentioned in the review of ed87e837, there could be a scenario where the
user sets the preferred size of an item inside the onWidthChanged handler
of another item:

    onWidthChanged: if (width < 10) secondItem.SplitView.preferredWidth = 100

Before this patch, this would result in the preferredWidth assignment being
ignored since it happened during a layout.

This patch adds some auto tests to ensure that this works, as the previous
patch (that converted layouts to be done in polish/updatePolish cycles)
already fixed the issue.

It also adds a check to avoid doing too many layouts in the case of
one of the split handles being dragged.

Change-Id: Ide519b33a2fa3bf746ae3793e0671fd1750c70d8
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>