When there are no points to grab, then we should return false
If there are no points to grab then it should not indicate that
it was able to grab points. Otherwise it can cause a handler to
be active when it has not valid points to work with. This can
happen if you do a pinch gesture on a DragHandler as the points
are not valid by default then.

Fixes: QTBUG-71972
Change-Id: Ib1c48e5a3777f4118944accfa2bc92edf0209884
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>